+++ KalturaPlugins/KalturaDocumentClientPlugin.php
@@ -660,7 +660,7 @@ class KalturaDocumentsService extends KalturaServiceBase
 			return $this->client->getMultiRequestResult();
 		$resultObject = $this->client->doQueue();
 		$this->client->throwExceptionIfError($resultObject);
-		$this->client->validateObjectType($resultObject, "bigint");
+		$this->client->validateObjectType($resultObject, "double");
 		return $resultObject;
 	}
